Split table clipboard tests.
